# Improve hashing performance