The improvement is in the range of ~10-20x depending
on platform and compilation mode, e.g. on my M1 MacBook Pro
AOT mode goes from 6ns/byte to 0.3ns/byte and JIT mode goes
from 5ns/byte to 0.4ns/byte.

The bulk of the improvement comes from avoiding allocating
a temporary ByteData for every readLE64. While it might
be reasonable to expect this allocation to be eliminated
Dart VM compiler is currently unable to do that for several
reasons (most of which can be easily fixed).

The second part of the win comes from avoiding boxed List
type for accumulator and using Int64List instead.

Additionally JIT compiler trips over `toUnsigned(64)` which
should be a no-op but instead hits a suboptimal implementation
of out of range left shift (1 << 64) and falls over fast path, so I
removed that as well.