@SamJakob 2024-09-28
Thanks! This is very interesting! I shall take a deeper dive and try and get this merged today.